015-registrando-e-excluindo-chaves-pix-no-bcb

+++ src/main/kotlin/br/com/zupedu/armando/pix/grpc/endpoints/RemoverChavePixEndpoint.kt
@@ -7,6 +7,8 @@ import br.com.zupedu.armando.core.handler.ErrorAroundHandler
 import br.com.zupedu.armando.core.handler.exceptions.ArgumentoDeEntradaInvalidoDefaultException
 import br.com.zupedu.armando.core.handler.exceptions.BadRequestErrorException
 import br.com.zupedu.armando.core.handler.exceptions.NotFoundDefaultException
+import br.com.zupedu.armando.httpclients.BcbClient
+import br.com.zupedu.armando.httpclients.DeletePixKeyRequest
 import br.com.zupedu.armando.httpclients.ItauErpClient
 import br.com.zupedu.armando.pix.grpc.extensions.toRemoverChavePixDto
 import br.com.zupedu.armando.pix.repository.ChavePixRepository
@@ -20,7 +22,8 @@ import javax.validation.Validator
 class RemoverChavePixEndpoint(
     private val validator: Validator,
     private val repository: ChavePixRepository,
-    private val itauErpClient: ItauErpClient
+    private val itauErpClient: ItauErpClient,
+    private val bcbClient: BcbClient
 ): PixKeyManagerRemoverServiceGrpc.PixKeyManagerRemoverServiceImplBase() {
     private val logger = LoggerFactory.getLogger(NovaChavePixEndpoint::class.java)
 
@@ -41,6 +44,8 @@ class RemoverChavePixEndpoint(
         logger.info("Validando se a chave pix pertence ao cliente informado")
         if (conta.titularCpf != chavePix.conta.titularCpf) throw BadRequestErrorException("ChavePix não pertence ao cliente informado.")
 
+        val deletarNoBcbResponse = bcbClient.deletarChavePix(chavePix.chave, DeletePixKeyRequest(chavePix.chave, chavePix.conta.instituicaoIspb))
+        if (deletarNoBcbResponse.status.code != 200) throw BadRequestErrorException("Ocorreu um erro ao deletar a chave pix no BCB.")
         repository.deleteById(chavePix.id)
         logger.info("Chave pix deletada com sucesso")
         responseObserver.onNext(RemoverPixResponse.newBuilder().setInfo("Chave Pix excluída com sucesso.").build())
